2019_06_27

+++ study/UnionFind.cpp
@@ -21,16 +21,16 @@ using Pll = pair<ll, ll>;
 #define MAX_N 100010
 #define MAX_Q 200010
 
-int N, Q;
-int P[MAX_Q], A[MAX_Q], B[MAX_Q], parent[MAX_N];
+ll N, Q;
+ll P[MAX_Q], A[MAX_Q], B[MAX_Q], parent[MAX_N];
 
-void init(int n) {
-    for (int i = 0; i < n; i++) {
+void init(ll n) {
+    for (ll i = 0; i < n; i++) {
         parent[i] = i;
     }
 }
 
-int root(int x) {
+ll root(ll x) {
     if (parent[x] == x) {
         return x;
     } else {
@@ -38,11 +38,11 @@ int root(int x) {
     }
 }
 
-bool same(int x, int y) {
+bool same(ll x, ll y) {
     return root(x) == root(y);
 }
 
-void unite(int x, int y) {
+void unite(ll x, ll y) {
     x = root(x);
     y = root(y);
 
@@ -57,13 +57,13 @@ int main(int argc, char *argv[])
 {
     cin >> N >> Q;
 
-    for (int i = 0; i < Q; i++) {
+    for (ll i = 0; i < Q; i++) {
         cin >> P[i] >> A[i] >> B[i];
     }
 
     init(N);
 
-    for (int i = 0; i < Q; i++) {
+    for (ll i = 0; i < Q; i++) {
         if (P[i]) {
             if (same(A[i], B[i])) {
                 Yes
